Major changes
+Added IP address entrance screen at the begining
+Orientation of the printing is flipped by x and y, so that it matches
with the lego printer.
+Error Animation is fixed by removing tooltip controller.
+The created bricks were appering without marker, hopefully they won't
appear anymore by looping trough all the renderers they have and
disabling them, when the marker is not found.
+When the server connection was lost in between the code had an
exception while trying to write the status on the screen. This exception
is handled, so the code will keep on working even if there is a
connection loss.